grasshopper driver returns image as type int. jkamgendriver now implements threading diffrently. KJmGenDriver is no longer moved to its own thread, this functionality as unused anyways, and FrameGrabber is now a QThread rather than a QObject being moved to a QThread.

+++ grasshopperdriver.py
@@ -80,7 +80,7 @@ class GrasshopperDriver(JKamGenDriver):
     def _grab_frame(cam):
         try:
             image_result = cam.GetNextImage(PySpin.EVENT_TIMEOUT_INFINITE)
-            frame = image_result.GetNDArray()
+            frame = image_result.GetNDArray().astype(int)
             frame = np.transpose(frame)
             image_result.Release()
             return frame
